# govctl_cli/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "SoloCorp OS — Admin API"
  description: |
    REST API for SoloCorp OS governance, agent orchestration, pipeline
    monitoring, and system health. Used by the Dashboard UI and external
    integrations.

    **Base URL:** `http://localhost:8765`

    **Authentication:** API key via `X-API-Key` header (required for write ops).
  version: "1.0.0"
  contact:
    name: "Software Architect (พี่ทรงศักดิ์)"
    url: "https://github.com/DrSoloDev/Lab-solocorp-os2.4"
servers:
  - url: http://localhost:8765
    description: "Local development server"
  - url: https://api.solocorp.local
    description: "Production server (future)"
paths:
...
  /api/v1/pipeline/status:
    get:
      operationId: get_pipeline_status
      summary: Get pipeline status for all or one project
      description: |
        Returns aggregated pipeline status including phase progress,
        guard status, queue depth, and recent events. If no project_id
        is specified, returns summary for all active projects.
      tags: [Pipeline]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
          description: "Filter to a single project (optional)"
          required: false
      responses:
        "200":
          description: "Pipeline status"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PipelineStatus"
                  - type: array
                    items:
                      $ref: "#/components/schemas/PipelineStatus"
  /api/v1/pipeline/history:
    get:
      operationId: get_pipeline_history
      summary: Get pipeline execution history
      description: |
        Returns timeline of pipeline events (phase transitions, guard
        checks, agent executions) for a project. Useful for debugging
        pipeline flow and auditing.
      tags: [Pipeline]
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
          description: "Project identifier"
        - name: since
          in: query
          schema:
            type: string
            format: date
          description: "Start date (YYYY-MM-DD). Default: today"
          required: false
        - name: until
          in: query
          schema:
            type: string
            format: date
          description: "End date (YYYY-MM-DD). Default: today"
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: "Maximum events to return"
          required: false
      responses:
        "200":
          description: "Pipeline history"
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PipelineEvent"
                  total:
                    type: integer
                  project_id:
                    type: string

  # ===================================================================
  # Monitoring
  # ===================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for write operations. Configurable in gov/config.toml"
  schemas:
...
    PipelineStatus:
      type: object
      required: [project_id, status, phase]
      properties:
        project_id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pending, in_progress, done, failed]
        phase:
          type: string
          enum: [spec, design, arch, dev, qa, deploy]
        progress_pct:
          type: integer
          minimum: 0
          maximum: 100
        phases:
          type: object
          description: "Per-phase status map"
          additionalProperties:
            type: object
            properties:
              status:
                type: string
                enum: [pending, in_progress, done, failed]
              owner:
                type: string
        blockers:
          type: array
          items:
            type: string
        guard_status:
          type: string
          enum: [pending, passed, failed]
        queue_depth:
          type: integer
          description: "Pending messages for this project"
        updated_at:
          type: string
          format: date-time
    PipelineEvent:
      type: object
      description: "A single pipeline event in the audit trail"
      required: [id, type, status, timestamp]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [governance, agent, deploy]
          description: "Event category"
        status:
          type: string
          enum: [pending, running, success, failed]
        source:
          type: string
          description: "Department ID that generated the event"
          example: "05-architect"
        timestamp:
          type: string
          format: date-time
        duration_ms:
          type: integer
          description: "Execution duration in milliseconds (if applicable)"
        detail:
          type: object
          description: "Event-specific payload"

    # ── Monitoring ───────────────────────────────────────────────────
...